restructure app
