+++ .gitignore
@@ -1,4 +1,2 @@
 data/**
-
-!data/categories.json
 !data/demo.yaml
